[ARP-86023] ARP IPF: Refactor ARP spec helpers.
