+++ modules/accredited_representative_portal/spec/controllers/concerns/accredited_representative_portal/authenticable_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
-require 'rails_helper'
-require_relative '../../../support/authentication'
+require_relative '../../../rails_helper'
 
 RSpec.describe AccreditedRepresentativePortal::Authenticable do
   controller(AccreditedRepresentativePortal::ApplicationController) do
