+++ modules/accredited_representative_portal/spec/requests/accredited_representative_portal/v0/user_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require_relative '../../../rails_helper'
-require_relative '../../../support/authentication'
 
 RSpec.describe 'AccreditedRepresentativePortal::V0::User', type: :request do
   describe '#show' do
